Make use of a header file
Extract the interface of common.c into a dedicated header, common.h,
and use that instead of including the C file.
