+++ src/run.combined.c
@@ -23,7 +23,7 @@
    authors and should not be interpreted as representing official policies, either expressed
    or implied, of Ivo Smits.*/
 
-#include "common.c"
+#include "common.h"
 
 extern struct qtproto qtproto_raw;
 extern struct qtproto qtproto_nacl0;
